% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.r
\name{cat_function}
\alias{cat_function}
\title{A Cat Function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE. If you put FALSE it calls you super not cool. GIt push}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
